Fix long standing bug in Vegas audio export
The end of audio export from Vegas has been having issues for a long
time. Users were adding a 1 second silence to work around it. Turns out
this was because we weren't checking for S_FALSE from the audio read
callback and reading wrong values from a buffer we had before.

https://github.com/satishsampath/frame-server/issues/45